+++ assets/javascripts/kitten/components/buttons/social-button-icon.test.js
@@ -29,9 +29,8 @@ describe('<FacebookButtonIcon />', () => {
   describe('with other props', () => {
     const button = shallow(
       <FacebookButtonIcon
-        className={ classNames('k-ButtonIcon__svg',
-                               'custom__class') }
-      />
+        className={classNames('k-ButtonIcon__svg', 'custom__class')}
+      />,
     )
 
     it('has a default classes', () => {
@@ -58,9 +57,8 @@ describe('<TwitterButtonIcon />', () => {
   describe('with other props', () => {
     const button = shallow(
       <TwitterButtonIcon
-        className={ classNames('k-ButtonIcon__svg',
-                               'custom__class') }
-      />
+        className={classNames('k-ButtonIcon__svg', 'custom__class')}
+      />,
     )
 
     it('has a default classes', () => {
@@ -87,9 +85,8 @@ describe('<LinkedinButtonIcon />', () => {
   describe('with other props', () => {
     const button = shallow(
       <LinkedinButtonIcon
-        className={ classNames('k-ButtonIcon__svg',
-                               'custom__class') }
-      />
+        className={classNames('k-ButtonIcon__svg', 'custom__class')}
+      />,
     )
 
     it('has a default classes', () => {
@@ -116,9 +113,8 @@ describe('<InstagramButtonIcon />', () => {
   describe('with other props', () => {
     const button = shallow(
       <InstagramButtonIcon
-        className={ classNames('k-ButtonIcon__svg',
-                               'custom__class') }
-      />
+        className={classNames('k-ButtonIcon__svg', 'custom__class')}
+      />,
     )
 
     it('has a default classes', () => {