+++ assets/javascripts/kitten/components/cards/crowdfunding-card.test.js
@@ -16,9 +16,7 @@ describe('<CrowdfundingCard />', () => {
 
   describe('by default', () => {
     beforeEach(() => {
-      component = renderer.create(
-        <CrowdfundingCard />
-      ).toJSON()
+      component = renderer.create(<CrowdfundingCard />).toJSON()
     })
 
     it('matches with snapshot', () => {
@@ -28,9 +26,7 @@ describe('<CrowdfundingCard />', () => {
 
   describe('with loading prop', () => {
     beforeEach(() => {
-      component = renderer.create(
-        <CrowdfundingCard loading />
-      ).toJSON()
+      component = renderer.create(<CrowdfundingCard loading />).toJSON()
     })
 
     it('matches with snapshot', () => {
@@ -40,34 +36,36 @@ describe('<CrowdfundingCard />', () => {
 
   describe('with some props', () => {
     beforeEach(() => {
-      component = renderer.create(
-        <CrowdfundingCard
-          href="#foobar"
-          imageProps={{
-            src: '#image',
-            alt: 'Image alt',
-            backgroundColor: '#d8d8d8',
-            color: '#333',
-          }}
-          avatarProps={{
-            src: '#avatar',
-            alt: 'Avatar alt',
-          }}
-          ownerTitle="Custom title"
-          ownerDescription="Custom description"
-          titleProps={{
-            tag: 'h4',
-          }}
-          cardTitle="Custom title"
-          cardSubTitle="Custom subtitle"
-          titlesMinHeight
-          info1="Custom information #1"
-          info2="Custom information #2"
-          info3="Custom information #3"
-          progress="84"
-          state="Custom state"
-        />
-      ).toJSON()
+      component = renderer
+        .create(
+          <CrowdfundingCard
+            href="#foobar"
+            imageProps={{
+              src: '#image',
+              alt: 'Image alt',
+              backgroundColor: '#d8d8d8',
+              color: '#333',
+            }}
+            avatarProps={{
+              src: '#avatar',
+              alt: 'Avatar alt',
+            }}
+            ownerTitle="Custom title"
+            ownerDescription="Custom description"
+            titleProps={{
+              tag: 'h4',
+            }}
+            cardTitle="Custom title"
+            cardSubTitle="Custom subtitle"
+            titlesMinHeight
+            info1="Custom information #1"
+            info2="Custom information #2"
+            info3="Custom information #3"
+            progress="84"
+            state="Custom state"
+          />,
+        )
+        .toJSON()
     })
 
     it('matches with snapshot', () => {