+++ assets/javascripts/kitten/components/cards/simple-card.test.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import renderer from 'react-test-renderer'
 import { SimpleCard } from 'kitten/components/cards/simple-card'
 import { Title } from 'kitten/components/typography/title'
 import { Text } from 'kitten/components/typography/text'
@@ -8,33 +9,12 @@ describe('<SimpleCard />', () => {
   let component
 
   describe('by default', () => {
-    beforeEach(() => {
-      component = mount(<SimpleCard />)
+    beforeAll(() => {
+      component = renderer.create(<SimpleCard />).toJSON()
     })
 
-    it('is a <div />', () => {
-      expect(component.render().is('div')).toBe(true)
-    })
-
-    it('has an image container', () => {
-      expect(component.render().find('.k-Card__imageContainer')).toHaveLength(1)
-    })
-
-    it('has a <img /> with good class', () => {
-      expect(component.find('img')).toHaveLength(1)
-      expect(component.find('img').hasClass('k-Card__image')).toBe(true)
-    })
-
-    it('has not a <Title /> component', () => {
-      expect(component.find(Title).exists()).toBeFalsy()
-    })
-
-    it('has not <Text /> component', () => {
-      expect(component.find(Text).exists()).toBeFalsy()
-    })
-
-    it('has <HorizontalStroke /> component', () => {
-      expect(component.find(HorizontalStroke).exists()).toBeTruthy()
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
     })
   })
 
@@ -56,14 +36,14 @@ describe('<SimpleCard />', () => {
     beforeEach(() => {
       component = mount(
         <SimpleCard
-          imageProps={ {
+          imageProps={{
             src: '#foobar',
             alt: 'FooBar',
             style: {
               cursor: 'crosshair',
             },
-          } }
-        />
+          }}
+        />,
       )
     })
 
@@ -103,7 +83,12 @@ describe('<SimpleCard />', () => {
     })
 
     it('has a <Text /> with good content', () => {
-      expect(component.find(Text).render().text()).toBe('Simple card subtitle')
+      expect(
+        component
+          .find(Text)
+          .render()
+          .text(),
+      ).toBe('Simple card subtitle')
       expect(component.find(Text).props()).toMatchObject({
         weight: 'regular',
         size: 'micro',
@@ -126,6 +111,16 @@ describe('<SimpleCard />', () => {
     })
   })
 
+  describe('without horizontalStroke prop', () => {
+    it('matches snapshot', () => {
+      component = renderer
+        .create(<SimpleCard horizontalStroke={false} />)
+        .toJSON()
+
+      expect(component).toMatchSnapshot()
+    })
+  })
+
   describe('with title props', () => {
     beforeEach(() => {
       component = mount(
@@ -135,7 +130,7 @@ describe('<SimpleCard />', () => {
             tag: 'h2',
             className: 'custom-class',
           }}
-        />
+        />,
       )
     })