+++ assets/javascripts/kitten/components/form/form-row.js
@@ -4,8 +4,10 @@ import classNames from 'classnames'
 export class FormRow extends React.Component {
   render() {
     return (
-      <div { ...this.props }
-           className={ classNames('k-FormRow', this.props.className) } />
+      <div
+        {...this.props}
+        className={classNames('k-FormRow', this.props.className)}
+      />
     )
   }
 }