+++ assets/javascripts/kitten/components/navigation/horizontal-nav.js
@@ -7,7 +7,7 @@ export class HorizontalNav extends Component {
   renderItems() {
     const { items } = this.props
 
-    return items.map((item) => this.renderItem(item))
+    return items.map(item => this.renderItem(item))
   }
 
   renderItem(item) {
@@ -21,11 +21,9 @@ export class HorizontalNav extends Component {
       ...others,
     } = item
 
-    const itemClassName = classNames(
-      'k-HorizontalNav__item',
-      className,
-      { 'is-selected': selected },
-    )
+    const itemClassName = classNames('k-HorizontalNav__item', className, {
+      'is-selected': selected,
+    })
 
     const renderBadge =
       <Fragment>
@@ -42,18 +40,20 @@ export class HorizontalNav extends Component {
 
     return (
       <li
-        className={ classNames('k-HorizontalNav__element',
-                               this.props.elementClassName) }
-        key={ key }
+        className={classNames(
+          'k-HorizontalNav__element',
+          this.props.elementClassName,
+        )}
+        key={key}
         role="menuitem"
-        { ...others }
+        {...others}
       >
         <a
-          href={ href }
-          className={ itemClassName }
-          style={ {
+          href={href}
+          className={itemClassName}
+          style={{
             height: this.props.height,
-          } }
+          }}
         >
           { renderTextAndBadge }
         </a>
@@ -62,25 +62,15 @@ export class HorizontalNav extends Component {
   }
 
   render() {
-    const {
-      className,
-      center,
-      ...others,
-    } = this.props
+    const { className, center, ...others } = this.props
 
-    const navClassNames = classNames(
-      'k-HorizontalNav',
-       className,
-      { 'k-HorizontalNav--center': center },
-    )
+    const navClassNames = classNames('k-HorizontalNav', className, {
+      'k-HorizontalNav--center': center,
+    })
 
     return (
-      <ul
-        role="menubar"
-        id={ this.props.id }
-        className={ navClassNames }
-      >
-        { this.renderItems() }
+      <ul role="menubar" id={this.props.id} className={navClassNames}>
+        {this.renderItems()}
       </ul>
     )
   }