+++ assets/javascripts/kitten/components/navigation/horizontal-nav.test.js
@@ -4,13 +4,13 @@ import { HorizontalNav } from 'kitten/components/navigation/horizontal-nav'
 describe('<HorizontalNav />', () => {
   const component = shallow(
     <HorizontalNav
-      items={ [
+      items={[
         { key: 'item-1', text: 'Nav link 1', href: '#foobar' },
         { key: 'item-2', text: 'Nav link 2', href: '#foobar' },
         { key: 'item-3', text: 'Nav link 3', href: '#foobar' },
         { key: 'item-4', text: 'Nav link 4', href: '#foobar' },
-      ] }
-    />
+      ]}
+    />,
   )
   const link = component.find('.k-HorizontalNav__item').first()
 
@@ -31,9 +31,7 @@ describe('<HorizontalNav />', () => {
   })
 
   describe('center props', () => {
-    const component = shallow(
-      <HorizontalNav center />
-    )
+    const component = shallow(<HorizontalNav center />)
 
     it('renders a className', () => {
       expect(component.hasClass('k-HorizontalNav--center')).toBe(true)
@@ -43,11 +41,9 @@ describe('<HorizontalNav />', () => {
   describe('height props', () => {
     const component = mount(
       <HorizontalNav
-        items={ [
-          { key: 'item-1', text: 'Nav link 1' },
-        ] }
+        items={[{ key: 'item-1', text: 'Nav link 1' }]}
         height="auto"
-      />
+      />,
     )
     const item = component.find('.k-HorizontalNav__item')
 
@@ -56,42 +52,44 @@ describe('<HorizontalNav />', () => {
     })
   })
 
-
   describe('with custom classes/id', () => {
     const component = shallow(
       <HorizontalNav
         id="custom-id"
         className="custom-class"
         elementClassName="element-custom-class"
-        items={ [
+        items={[
           { key: 'item-1', text: 'Nav link 1', className: 'item-custom-class' },
           { key: 'item-2', text: 'Nav link 2', className: 'item-custom-class' },
           { key: 'item-3', text: 'Nav link 3', className: 'item-custom-class' },
           { key: 'item-4', text: 'Nav link 4', className: 'item-custom-class' },
-        ] }
-      />
+        ]}
+      />,
     )
 
     it('renders the right classes', () => {
       expect(component.props().id).toBe('custom-id')
       expect(component.hasClass('custom-class')).toBe(true)
       expect(
-        component.find('.k-HorizontalNav__element').first()
-          .hasClass('element-custom-class')
+        component
+          .find('.k-HorizontalNav__element')
+          .first()
+          .hasClass('element-custom-class'),
       ).toBe(true)
 
-      expect(component.find('.k-HorizontalNav__item').first()
-        .hasClass('item-custom-class')
+      expect(
+        component
+          .find('.k-HorizontalNav__item')
+          .first()
+          .hasClass('item-custom-class'),
       ).toBe(true)
     })
 
     describe('with HTML in items', () => {
       const component = shallow(
         <HorizontalNav
-          items={ [
-            { key: 'item-1', text: 'Nav\n1', className: 'item-1' },
-          ] }
-        />
+          items={[{ key: 'item-1', text: 'Nav\n1', className: 'item-1' }]}
+        />,
       )
 
       it('transforms line break with <br/>', () => {