+++ assets/javascripts/kitten/karl/layout/line.js
@@ -8,13 +8,10 @@ import { HeartIcon } from 'kitten/components/icons/heart-icon'
 export const KarlLineWithTooltip = props => {
   return (
     <Line>
+      <Line.Item>{props.text}</Line.Item>
       <Line.Item>
-        { props.text }
-      </Line.Item>
-      <Line.Item>
-        <Tooltip id={ props.id }
-                 place={ props.place }>
-          { props.children }
+        <Tooltip id={props.id} place={props.place}>
+          {props.children}
         </Tooltip>
       </Line.Item>
     </Line>
@@ -24,9 +21,7 @@ export const KarlLineWithTooltip = props => {
 export const KarlLineWithBigHeart = props => {
   return (
     <Line>
-      <Line.Item>
-        Lorem ipsum dolor sit amet.
-      </Line.Item>
+      <Line.Item>Lorem ipsum dolor sit amet.</Line.Item>
       <Line.Item>
         <button className="k-ButtonIcon k-ButtonIcon--hydrogen k-ButtonIcon--big">
           <HeartIcon className="k-ButtonIcon__svg" />