+++ CHANGELOG.md
@@ -5,7 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `Badge` prop on `HorizontalNav` component.
+- Fix: Update `:hover` and `:focus` state on `azote` button.
 - Feature: Add `subTitleTruncate` prop for `CrowdfundingCard`.
+- Feature: Add `horizontalStroke` prop for `SimpleCard`.
 
 ## [20.5.0] - 2018-04-03